Move DeviceInputSystem to its own file
