+++ Plugins/NativeInput/Include/Babylon/Plugins/NativeInput.h
@@ -4,11 +4,12 @@
 
 namespace Babylon::Plugins
 {
-    class NativeInput
+    class NativeInput final
     {
     public:
         // TODO: Ideally instances of these should be scoped to individual views within an env, but we don't yet support multi-view.
         static NativeInput& CreateForJavaScript(Napi::Env);
+        static NativeInput& GetFromJavaScript(Napi::Env);
 
         void PointerDown(uint32_t pointerId, uint32_t buttonIndex, uint32_t x, uint32_t y);
         void PointerUp(uint32_t pointerId, uint32_t buttonIndex, uint32_t x, uint32_t y);
@@ -22,7 +23,5 @@ namespace Babylon::Plugins
         NativeInput(Napi::Env);
         class Impl;
         std::unique_ptr<Impl> m_impl{};
-
-    private:
     };
 }
